backend/requirements.txt
asgiref~=3.5.0
Django~=4.0.2
django-autoslug~=1.9.8
django-cors-headers~=3.11.0
django-extensions~=3.1.5
django-nested-admin~=3.4.0
django-templated-mail~=1.1.1
djangorestframework~=3.13.1
djangorestframework-simplejwt~=4.8.0
djoser~=2.1.0
drf-spectacular~=0.21.2
drf-writable-nested~=0.6.3
pytz~=2021.3
sqlparse~=0.4.2
django_fsm~=2.8.0
pandas
matplotlib
numpy
channels~=3.0.4
django-eventstream~=4.4.1
celery
django-celery-email
# elasticsearch-dsl==7.0.0
git+https://github.com/kisiel/django-elasticsearch-dsl-drf.git@a27fa1eb7739c8f989b15f7ef891969be3105f68